console: Move validation to location form
